修复 stat fstat 系统调用错误返回值, 修复 read write 没有文件读取偏移问题, 修复文件错误 inode 计数问题(临时解决, 可运行 rm ls 命令)

+++ src/x86_64/driver/vdisk.c
@@ -24,7 +24,7 @@ bool have_vdisk(int drive) {
     }
 }
 
-uint32_t disk_size(int drive) {
+size_t disk_size(int drive) {
     uint8_t drive1 = drive;
     if (have_vdisk(drive1)) {
         int indx = drive1;
